component folder refactored
