+++ b2b-sample/guardio-insurance-sample-application/components/common/dialog.js
