+++ b2b-sample/guardio-insurance-sample-application/components/sections/settingsSection/idpSection/data/templates/google.json
