+++ b2b-sample/guardio-insurance-sample-application/components/sections/settingsSection/idpSection/otherComponents/idpDetailsSections/general.js
@@ -25,8 +25,8 @@ import decodePatchGeneralSettingsIdp from
     "../../../../../../util/apiDecode/settings/identityProvider/decodePatchGeneralSettingsIdp";
 import { checkIfJSONisEmpty } from "../../../../../../util/util/common/common";
 import { LOADING_DISPLAY_BLOCK, LOADING_DISPLAY_NONE } from "../../../../../../util/util/frontendUtil/frontendUtil";
-import { errorTypeDialog, successTypeDialog } from "../../../../../util/dialog";
-import HelperText from "../../../../../util/helperText";
+import { errorTypeDialog, successTypeDialog } from "../../../../../common/dialog";
+import HelperText from "../../../../../common/helperText";
 
 /**
  *
