+++ b2b-sample/guardio-insurance-sample-application/components/sections/settingsSection/manageUserSection/manageUserSectionComponent.js
@@ -22,7 +22,7 @@ import AddUserComponent from "./otherComponents/addUserComponent";
 import styles from "../../../../styles/Settings.module.css";
 import decodeViewUsers from "../../../../util/apiDecode/settings/decodeViewUsers";
 import EditUserComponent from "./otherComponents/editUserComponent";
-import SettingsTitle from "../../../util/settingsTitle";
+import SettingsTitle from "../../../common/settingsTitle";
 
 /**
  *
